Fix spinoff bearing more

+++ common/src/main/kotlin/org/valkyrienskies/clockwork/content/forces/contraption/BearingController.kt
@@ -235,7 +235,7 @@ class BearingController : ShipPhysicsListener {
             return Vector3d()
         }
 
-        val maxTorque = ClockworkConfig.SERVER.angleFollowingMaxTorque
+        val maxTorque = ClockworkConfig.SERVER.angleFollowingMaxTorque.toDouble()
         if (maxTorque > 0.0) {
             torque = torque.coerceIn(-maxTorque, maxTorque)
         }
